HLDN-647 Conform enum generation to project style
