+++ extensions/openapi-generator/modules/openapi-generator/src/main/resources/helidon-declarative/stringEnum.mustache
@@ -1,9 +1,24 @@
+{{!
+Copyright (c) 2026 Oracle and/or its affiliates.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+}}
+
 @Json.Converter({{converterName}}.class)
 public enum {{enumName}} {
 {{#values}}
     {{name}}({{{value}}}){{^-last}},{{/-last}}{{#-last}};{{/-last}}
 {{/values}}
-
     private final String value;
 
     {{enumName}}(String value) {
